cmake: add some simple tests on windows platforms.
Our main test harnesses on written in unix shell script, so
the ctest wasn't working on windows, and a stanze was inserted
to skip all tests to get a green build.

However, this is misleading since no testing at all was done.
Instead, build a separate set of tests on non-unix platforms,
calling the two api test programs, which of necessity are
compiled executables. Also invoke the two front-end programs
with just the --help switch so verify they at least start.

Thanks to Marcus Asteborg for the suggestion.